+++ src/Symfony/Component/Mailer/Bridge/Amazon/Tests/Transport/SesApiAsyncAwsTransportTest.php
@@ -21,7 +21,9 @@ use Symfony\Component\HttpClient\Response\MockResponse;
 use Symfony\Component\Mailer\Bridge\Amazon\Transport\SesApiAsyncAwsTransport;
 use Symfony\Component\Mailer\Envelope;
 use Symfony\Component\Mailer\Exception\HttpTransportException;
+use Symfony\Component\Mailer\Exception\InvalidArgumentException;
 use Symfony\Component\Mailer\Header\MetadataHeader;
+use Symfony\Component\Mailer\RemoteTemplateEmail;
 use Symfony\Component\Mime\Address;
 use Symfony\Component\Mime\Email;
 use Symfony\Component\Mime\RawMessage;
@@ -191,6 +193,71 @@ class SesApiAsyncAwsTransportTest extends TestCase
         $message = $transport->send($mail);
 
         $this->assertSame('foobar', $message->getMessageId());
+    }
+
+    public function testSendRemoteTemplate()
+    {
+        $client = new MockHttpClient(function (string $method, string $url, array $options): ResponseInterface {
+            $this->assertSame('POST', $method);
+            $this->assertSame('https://email.us-east-1.amazonaws.com/v2/email/outbound-emails', $url);
+
+            $content = json_decode($options['body'], true);
+
+            $this->assertSame('welcome', $content['Content']['Template']['TemplateName']);
+            $this->assertSame('{"firstName":"Fabien"}', $content['Content']['Template']['TemplateData']);
+            $this->assertSame([['Name' => 'X-Custom-Header', 'Value' => 'foobar']], $content['Content']['Template']['Headers']);
+            $this->assertArrayNotHasKey('Simple', $content['Content']);
+
+            $json = '{"MessageId": "foobar"}';
+
+            return new MockResponse($json, [
+                'http_code' => 200,
+            ]);
+        });
+
+        $transport = new SesApiAsyncAwsTransport(new SesClient(Configuration::create(['sharedConfigFile' => false]), new NullProvider(), $client));
+
+        $mail = (new RemoteTemplateEmail())
+            ->to(new Address('saif.gmati@symfony.com', 'Saif Eddin'))
+            ->from(new Address('fabpot@symfony.com', 'Fabien'))
+            ->template('welcome', ['firstName' => 'Fabien']);
+        $mail->getHeaders()->addTextHeader('X-Custom-Header', 'foobar');
+
+        $message = $transport->send($mail);
+
+        $this->assertSame('foobar', $message->getMessageId());
+    }
+
+    public function testRemoteTemplateRejectsSubject()
+    {
+        $transport = new SesApiAsyncAwsTransport(new SesClient(Configuration::create(['sharedConfigFile' => false]), new NullProvider(), new MockHttpClient()));
+
+        $mail = (new RemoteTemplateEmail())
+            ->to(new Address('saif.gmati@symfony.com', 'Saif Eddin'))
+            ->from(new Address('fabpot@symfony.com', 'Fabien'))
+            ->subject('Hello!')
+            ->template('welcome');
+
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage('Amazon SES does not support overriding the subject of a template; define the subject in the template itself.');
+
+        $transport->send($mail);
+    }
+
+    public function testRemoteTemplateRejectsAttachments()
+    {
+        $transport = new SesApiAsyncAwsTransport(new SesClient(Configuration::create(['sharedConfigFile' => false]), new NullProvider(), new MockHttpClient()));
+
+        $mail = (new RemoteTemplateEmail())
+            ->to(new Address('saif.gmati@symfony.com', 'Saif Eddin'))
+            ->from(new Address('fabpot@symfony.com', 'Fabien'))
+            ->template('welcome')
+            ->attach('some attachment', 'attachment.txt');
+
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage('The Amazon SES API does not support attachments when using a remote template.');
+
+        $transport->send($mail);
     }
 
     public function testSendThrowsForErrorResponse()