+++ src/Symfony/Component/Mailer/Bridge/Mailjet/Transport/MailjetApiTransport.php
@@ -15,10 +15,13 @@ use Psr\EventDispatcher\EventDispatcherInterface;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\Mailer\Envelope;
 use Symfony\Component\Mailer\Exception\HttpTransportException;
+use Symfony\Component\Mailer\Exception\InvalidArgumentException;
 use Symfony\Component\Mailer\Exception\TransportException;
 use Symfony\Component\Mailer\Header\TrackingHeader;
+use Symfony\Component\Mailer\RemoteTemplateEmail;
 use Symfony\Component\Mailer\SentMessage;
 use Symfony\Component\Mailer\Transport\AbstractApiTransport;
+use Symfony\Component\Mailer\Transport\RemoteTemplateTransportInterface;
 use Symfony\Component\Mime\Address;
 use Symfony\Component\Mime\Email;
 use Symfony\Contracts\HttpClient\Exception\DecodingExceptionInterface;
@@ -26,7 +29,7 @@ use Symfony\Contracts\HttpClient\Exception\TransportExceptionInterface;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 use Symfony\Contracts\HttpClient\ResponseInterface;
 
-class MailjetApiTransport extends AbstractApiTransport
+class MailjetApiTransport extends AbstractApiTransport implements RemoteTemplateTransportInterface
 {
     private const HOST = 'api.mailjet.com';
     private const API_VERSION = '3.1';
@@ -113,14 +116,17 @@ class MailjetApiTransport extends AbstractApiTransport
             $html = stream_get_contents($html);
         }
         [$attachments, $inlines, $html] = $this->prepareAttachments($email, $html);
+        $template = $email instanceof RemoteTemplateEmail ? $email->getRemoteTemplate() : null;
 
         $message = [
             'From' => $this->formatAddress($envelope->getSender()),
             'To' => $this->formatAddresses($this->getRecipients($email, $envelope)),
-            'Subject' => $email->getSubject(),
-            'Attachments' => $attachments,
-            'InlinedAttachments' => $inlines,
         ];
+        if (null === $template || null !== $email->getSubject()) {
+            $message['Subject'] = $email->getSubject();
+        }
+        $message['Attachments'] = $attachments;
+        $message['InlinedAttachments'] = $inlines;
         if ($emails = $email->getCc()) {
             $message['Cc'] = $this->formatAddresses($emails);
         }
@@ -152,6 +158,9 @@ class MailjetApiTransport extends AbstractApiTransport
 
         foreach ($email->getHeaders()->all() as $headerName => $header) {
             if ($convertConf = self::HEADER_TO_MESSAGE[$headerName] ?? false) {
+                if ('x-mj-templateid' === $headerName) {
+                    trigger_deprecation('symfony/mailjet-mailer', '8.2', 'Using the "X-MJ-TemplateID" email header to select a Mailjet template is deprecated, use a "%s" instead.', RemoteTemplateEmail::class);
+                }
                 $message[$convertConf[0]] = $this->castCustomHeader($header->getBodyAsString(), $convertConf[1]);
                 continue;
             }
@@ -165,6 +174,17 @@ class MailjetApiTransport extends AbstractApiTransport
             }
 
             $message['Headers'][$header->getName()] = $header->getBodyAsString();
+        }
+
+        if (null !== $template) {
+            if (!ctype_digit($template->getReference())) {
+                throw new InvalidArgumentException(\sprintf('The Mailjet API expects a numeric template id, "%s" given.', $template->getReference()));
+            }
+            $message['TemplateID'] = (int) $template->getReference();
+            $message['TemplateLanguage'] = true;
+            if ($template->getVariables()) {
+                $message['Variables'] = $template->getVariables();
+            }
         }
 
         return [