+++ src/Symfony/Component/Mailer/Bridge/MailerSend/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 8.2
 ---
 
+ * Add support for `RemoteTemplateEmail` to `MailerSendApiTransport`
  * Add support for `TagHeader` in the API and SMTP transports
 
 7.1