+++ src/Symfony/Component/Mailer/Bridge/Mailjet/CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+8.2
+---
+
+ * Add support for `RemoteTemplateEmail` to `MailjetApiTransport`
+ * Deprecate the "X-MJ-TemplateID" email header, use a `RemoteTemplateEmail` instead
+
 6.4
 ---